refactor jni part
