+++ concordium-sdk/src/test/java/com/concordium/sdk/crypto/ed25519/NativeResolverTest.java
@@ -1,5 +1,6 @@
 package com.concordium.sdk.crypto.ed25519;
 
+import com.concordium.sdk.crypto.NativeResolver;
 import org.junit.Test;
 
 public class NativeResolverTest {
