--- concordium-sdk/src/test/java/com/concordium/sdk/crypto/encryptedtransfers/NativeResolverTest.java
@@ -1,11 +1,0 @@
-package com.concordium.sdk.crypto.encryptedtransfers;
-
-import org.junit.Test;
-
-public class NativeResolverTest {
-
-    @Test
-    public void testCanLoadNativeLib (){
-        NativeResolver.loadLib();
-    }
-}
