+++ crypto-jni/Cargo.toml
@@ -1,5 +1,5 @@
 [package]
-name = "ed25519-jni"
+name = "crypto-jni"
 version = "0.1.0"
 edition = "2018"
